% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Cleans source data from NOAA}
\usage{
eq_clean_data(data)
}
\arguments{
\item{data}{A data frame with raw data obtained from NOAA website}
}
\value{
A data frame with cleaned date, latitude, longitude and location
columns
}
\description{
\code{eq_clean_data} cleans date, latitude and longitude, and location name
from the source NOAA data
}
\details{
The function adds a column DATE with cleaned date, transforms LATITUDE and
LONGITUDE columns as numeric objects and transforms LOCATION_NAME by removing
the country and transforming to title case.
}
\examples{
\dontrun{
data <- readr::read_delim("earthquakes.tsv.gz", delim = "\\t")
clean_data <- eq_clean_data(data)
}

}

