% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_element.R
\name{theme_timeline}
\alias{theme_timeline}
\title{Theme for the timeline visualisation in ggplot2}
\usage{
theme_timeline()
}
\description{
Timeline visualisation
}
\examples{
\dontrun{
data \%>\% eq_clean_data() \%>\%
filter(COUNTRY \%in\% c("GREECE", "ITALY"), YEAR > 2000) \%>\%
   ggplot(aes(x = DATE,
              y = COUNTRY,
              color = as.numeric(TOTAL_DEATHS),
              size = as.numeric(EQ_PRIMARY)
   )) +
   geom_timeline() +
   theme_timeline()
}

}

